move lane connection manager to sub folder.
