+++ TLM/TLM/Util/PriorityRoad.cs
@@ -6,6 +6,7 @@ namespace TrafficManager.Util {
     using TrafficManager.API.Traffic.Data;
     using TrafficManager.API.Traffic.Enums;
     using TrafficManager.Manager.Impl;
+    using TrafficManager.Manager.Impl.LaneConnection;
     using TrafficManager.State;
     using TrafficManager.UI.SubTools.PrioritySigns;
     using TrafficManager.Util.Extensions;
